add remaining subcommands, rename publish to deploy
