+++ src/subcmds/mod.rs
@@ -1,6 +1,6 @@
 pub mod build;
 pub mod clean;
+pub mod deploy;
 pub mod module;
 pub mod new;
-pub mod publish;
 pub mod run;
